requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
django-environ>=0.11
django-cors-headers>=4.4
psycopg2-binary>=2.9 ; platform_system!="Windows"
psycopg2>=2.9 ; platform_system=="Windows"
pytest>=8.2
pytest-django>=4.8
pytest-cov>=5.0
model-bakery>=1.18
freezegun>=1.5